MODIFY: organize folder & fix scroll event
