+++ src/app/globals.css
@@ -2,8 +2,9 @@
 @tailwind components;
 @tailwind utilities;
 
-html,
+html{
+  overflow-y: auto;
+}
 body {
   height: 100vh;
-  overflow-y: auto;
 }
