+++ src/app/layout.tsx
@@ -1,7 +1,7 @@
-import Header from "@/components/Header";
+import Header from "@/components/common/Header";
 import "./globals.css";
 import { Open_Sans } from "next/font/google";
-import Footer from "@/components/Footer";
+import Footer from "@/components/common/Footer";
 import { Metadata } from "next";
 import AuthContext from "./context/AuthContext";
