+++ src/components/common/Footer.tsx
@@ -1,6 +1,7 @@
 'use client'
 
 import { RxPinTop } from "react-icons/rx";
+import React from 'react';
 
 export default function Footer() {
     const scrollToTop = () => {
@@ -12,13 +13,12 @@ export default function Footer() {
 
     return (
         <div>
-            <a
-                href=""
-                onClick={() => scrollToTop()}
+            <button
+                onClick={scrollToTop}
                 className="fixed bottom-5 right-7 shadow-2xl border border-gray-400 bg-white text-gray-400 p-2 rounded-full hover:bg-gray-100"
             >
                 <RxPinTop size={30} />
-            </a>
+            </button>
             <p className="bg-gray-700 text-white py-2 text-sm text-center">{"You can look back but don't regret it | All Right Reserved by Giyoun"}</p>
         </div>
     );
